Move cmd/system to subpackage
