+++ cmd/nerdctl/container/multi_platform_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package container
 
 import (
 	"fmt"
