+++ cmd/nerdctl/system/system_events_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package system
 
 import (
 	"fmt"
