+++ cmd/nerdctl/system/system_prune.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package system
 
 import (
 	"fmt"
